Voltage.C methods to check all aspects of open wire

+++ BeVolt/Drivers/inc/LTC681x.h
@@ -491,7 +491,7 @@ int16_t LTC681x_run_adc_redundancy_st(uint8_t adc_mode,
                                       cell_asic ic[]);
 
 /*! Helper function that runs the datasheet open wire algorithm*/
-long LTC681x_run_openwire_multi(uint8_t total_ic,
+uint32_t LTC681x_run_openwire_multi(uint8_t total_ic,
                           cell_asic ic[],
 													bool print);
